requirements.txt
azure-core==1.26.1
azure-storage-blob==12.14.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
msrest==0.7.1
numpy==1.23.5
oauthlib==3.2.2
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.9.3
six==1.16.0
SQLAlchemy==1.4.45
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
xlrd==2.0.1
